spapr/xive: add migration support for KVM
This extends the KVM XIVE models to handle the state synchronization
with KVM, for the monitor usage and for the migration.

Migration requires a VM change state handler to mark the EQ pages
dirty when the XIVE IC is fully stabilized. When the VM is stopped,
the handler masks the sources (PQ=01) to stop the flow of events and
saves the previous state. The XIVE controller is then synched through
KVM to flush any in-flight event notification and stabilize the
EQs. At this stage, the EQ page can be marked dirty to let the
migration sequence transfer the EQ pages to the destination, which is
done just after the stop state.

The XIVE device vmstates are then all captured by the sPAPRXive
interrupt controller model. Its priority is raised for KVM to operate
first and to capture the XIVE EAT and ENDT tables and the thread
interrupt contexts.

At restart, the sPAPRXive 'post_load' method is called by the sPAPR
machine and not by the migration engine. This to be sure that the KVM
XIVE device is created and after all XIVE device vmstates have been
transfered and loaded. First, sPAPRXive restores the XIVE routing
tables: ENDT and EAT. Next, are restored the thread interrupt context
registers. Finally, the sources configuration is restored in the VM
change state handler when the machine reaches the running state.

The get/set operations rely on their KVM counterpart in the host
kernel which acts as a proxy for OPAL, the host firmware.

Signed-off-by: Cédric Le Goater <clg@kaod.org>