spapr: add a 'reset' method to the sPAPR IRQ backend
This method will become useful when the new machine supporting both
interrupt modes, XIVE and XICS, is introduced. In this machine, the
interrupt mode is chosen by the CAS negotiation process and activated
after a reset.

For the time being, the only thing that can be done in the XIVE reset
handler is to map the pages for the TIMA and for the source ESBs.

Signed-off-by: Cédric Le Goater <clg@kaod.org>