target/ppc: Fix support for "STOP light" states on POWER9
STOP must act differently based on PSSCR:EC on POWER9. When
set, it acts like the P7/P8 power management instructions and
wake up at 0x100 based on the wakeup conditions in LPCR.

When PSSCR:EC is clear however it will wakeup at the next
instruction after STOP (if EE is clear) or take the corresponding
interrupts (if EE is set).

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Signed-off-by: Cédric Le Goater <clg@kaod.org>