ppc/pnv: fake a BMC address space
This populates the AHB address space of the SuperIO iLPC2AHB bridge
device with the Aspeed SoC controllers registers needed to drive the
PNOR.

Instantiating an AspeedSoc object on this address space would be
preferable but this still would be a Frankenstein machine mixing PPC
and ARM models. Nevertheless, it is a required step to give PNOR
support to PowerNV machine.

To go further, we would need a better LPC model on both side, QEMU
support for hybrid architectures, etc.

Signed-off-by: Cédric Le Goater <clg@kaod.org>