aspeed/i2c: add support for DMA transfers
The Aspeed AST2500 SoC adds a new feature to I2C controller which
enables DMA transfers to and from DRAM.

A pair of registers defines the buffer address and the length of the
DMA transfer. The address should be aligned on 4 bytes and the maximum
length should not exceed 4K. The receive or transmit DMA transfer can
then be initiated with specific bits in the Command/Status register of
the controller.

Needs OS support to test.

Signed-off-by: Cédric Le Goater <clg@kaod.org>